FM wideband rad1o
This relies on the filter code from portalib.

On the other hand, it droppes the GPDMA based fetching of sample
data from the SGPIO as used by portalib: The samples were heavily
distorted.

There seems to be a bug in the single-plane SGPIO mode, maybe
it's the CPLD interface or its firmware.

So we use the eight-plane SGPIO interface that is also used
by the hackrf USB firmware. It makes timing a bit of a hassle,
but it works quite well.

Will sample 12.288 MHz on the ADC, drop 3 of 4 samples in the CPLD,
and downsample to 768 kHz, do FM demodulation, then downsample to
192 kHz and that's our sample rate on the rad1o's DAC.

Sounds quite nice in the end. You'll want a nice antenna - d1pole
works well.