FM wideband rad1o: Transmit functionality
This patch adds transmit functionality to the FM wideband radio
applet.

You can switch between receiving and transmitting either PTT style
(chose RX/TX menu item, hold stick to the right to switch to TX,
release to switch back to RX) or in a back-and-forth manner (press
stick to the left and release to toggle between RX and TX).