rfapp: generate new cos/sin LUT, provide generation script, fix order
Also, switch the order of sin() and cos() values. Also, restrict
accessors to a single complex_s8_t one, so the naming "cos_sin"
now actually refers to the complex numbers: The Q value follows
a cosine, the I value a sine (like in most SDR examples and graphs
you will see in literature). As the complex_s8_t type stores the
values with I in the first octet, Q in the second octet, we had
the values stored in the wrong order before.