rflib: isolate sample handling on m0 core
This approach loads of handling of RF sample data to the m0 core.

Currently, TX is implemented to the point that the M0 core can
generate BFSK modulated sample data for a data buffer.

RX is more congested due to the limited instruction set of the M0
core. Currently, the core can do -fs/4 frequency translation plus
two-stage 1-3-3-1 FIR filtering plus decimation, resulting in an
output sample rate of fs/4, gain of 8. This works up to a sample
rate of 2500000.

Alternatively, an RX path is implemented where the M0 core calculates
atan2(I,Q) over the samples after the same preprocessing as mentioned
above. However, this only works for sample rates up to 1500000.

m0 subfolder contains the M0 code and a linker script. The m0rxtx.h
is supposed to be used by both the M0 and the M4 code.