rflib: actual library
rflib is now really a library, bfsk.c is just the demo program to
showcase it.

All (!) RF related operations are done on the M0 core, including
setup/shutdown of the RF parts. All that is left is a tiny shiny
API.

M0 code is currently about 12kByte and still linked into the M4 code.
This ought to be changed so it is loaded from a file directly into
memory.