rfapp: BTLE advertisement scanner
A simple GFSK demodulator and BT LE decoder. Will output MAC
address and name of BT LE devices that send advertisements.
For a simple app on Android that does send those, see e.g.
https://play.google.com/store/apps/details?id=com.nehori.java_eddystone_url_sender_androidtv
For the Android API, see
http://developer.android.com/reference/android/bluetooth/le/BluetoothLeAdvertiser.html
For Linux and the Bluez stack, see e.g.
https://github.com/google/eddystone/tree/master/eddystone-url/implementations/linux-url-advertiser
All of these send advertisements.

Could be enhanced quite a bit to do something meaningful with
other data as well. Also, this is RX only for now, TX is yet to
be done.